# docs/api/openapi.json
{
 "components": {
  "schemas": {
...
   "DepartmentCreate": {
    "additionalProperties": false,
    "properties": {
     "code": {
      "maxLength": 100,
      "minLength": 1,
      "title": "Code",
      "type": "string"
     },
     "name": {
      "maxLength": 200,
      "minLength": 1,
      "title": "Name",
      "type": "string"
     }
    },
    "required": [
     "code",
     "name"
    ],
    "title": "DepartmentCreate",
    "type": "object"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "RoleCreate": {
    "additionalProperties": false,
    "properties": {
     "code": {
      "maxLength": 100,
      "minLength": 1,
      "title": "Code",
      "type": "string"
     },
     "name": {
      "maxLength": 200,
      "minLength": 1,
      "title": "Name",
      "type": "string"
     }
    },
    "required": [
     "code",
     "name"
    ],
    "title": "RoleCreate",
    "type": "object"
   },
...
   "RoleUpdate": {
    "additionalProperties": false,
    "properties": {
     "name": {
      "anyOf": [
       {
        "maxLength": 200,
        "minLength": 1,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Name"
     },
     "permissions": {
      "anyOf": [
       {
        "items": {
         "type": "string"
        },
        "maxItems": 200,
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Permissions"
     }
    },
    "title": "RoleUpdate",
    "type": "object"
   },
...
   "UserCreate": {
    "additionalProperties": false,
    "description": "47.1.3 account resolution r3 \u2014 LegalMind does not self-provision. An\naccount exists only because an authorized administrator created it, so its\nroles are always assigned deliberately and never inferred from a login.",
    "properties": {
     "email": {
      "maxLength": 320,
      "minLength": 3,
      "title": "Email",
      "type": "string"
     },
     "name": {
      "maxLength": 200,
      "minLength": 1,
      "title": "Name",
      "type": "string"
     }
    },
    "required": [
     "email",
     "name"
    ],
    "title": "UserCreate",
    "type": "object"
   },
   "UserStatus": {
    "description": "42.2.",
    "enum": [
     "ACTIVE",
     "SUSPENDED",
     "DISABLED"
    ],
    "title": "UserStatus",
    "type": "string"
   },
...
  }
 },
 "info": {
  "title": "LegalMind API",
  "version": "1"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/v1/departments": {
   "get": {
    "operationId": "list_departments_api_v1_departments_get",
    "parameters": [
     {
      "in": "query",
      "name": "page",
      "required": false,
      "schema": {
       "default": 1,
       "minimum": 1,
       "title": "Page",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "page_size",
      "required": false,
      "schema": {
       "default": 25,
       "minimum": 1,
       "title": "Page Size",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response List Departments Api V1 Departments Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "List Departments",
    "tags": [
     "administration"
    ]
   },
   "post": {
    "description": "A department is created empty. Membership is set per user through\n``PATCH /users/{id}`` \u2014 one audited act per person, never a bulk move.",
    "operationId": "create_department_api_v1_departments_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/DepartmentCreate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Create Department Api V1 Departments Post",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Create Department",
    "tags": [
     "administration"
    ]
   }
  },
...
  "/api/v1/roles": {
   "get": {
    "operationId": "list_roles_api_v1_roles_get",
    "parameters": [
     {
      "in": "query",
      "name": "page",
      "required": false,
      "schema": {
       "default": 1,
       "minimum": 1,
       "title": "Page",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "page_size",
      "required": false,
      "schema": {
       "default": 25,
       "minimum": 1,
       "title": "Page Size",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response List Roles Api V1 Roles Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "List Roles",
    "tags": [
     "administration"
    ]
   },
   "post": {
    "description": "A new role starts with **no permissions**. Granting them is a separate,\nS-8-guarded act.",
    "operationId": "create_role_api_v1_roles_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RoleCreate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Create Role Api V1 Roles Post",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Create Role",
    "tags": [
     "administration"
    ]
   }
  },
  "/api/v1/roles/{role_id}": {
   "patch": {
    "description": "Replace a role's permission set \u2014 S-10, locked 43.26.\n\nThe replacement is transactional: the whole set changes or none of it does. The\nexternal reference's non-transactional delete-then-insert is rejected (C-EXT-7)\nbecause a failure between the two leaves a role with *no* permissions, which is\na silent authority change.",
    "operationId": "update_role_api_v1_roles__role_id__patch",
    "parameters": [
     {
      "in": "path",
      "name": "role_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Role Id",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RoleUpdate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Update Role Api V1 Roles  Role Id  Patch",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Update Role",
    "tags": [
     "administration"
    ]
   }
  },
  "/api/v1/users": {
   "get": {
    "description": "List users with optional filtering by status and search by email/name.",
    "operationId": "list_users_api_v1_users_get",
    "parameters": [
     {
      "in": "query",
      "name": "status",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "$ref": "#/components/schemas/UserStatus"
        },
        {
         "type": "null"
        }
       ],
       "title": "Status"
      }
     },
     {
      "in": "query",
      "name": "search",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Search"
      }
     },
     {
      "in": "query",
      "name": "page",
      "required": false,
      "schema": {
       "default": 1,
       "minimum": 1,
       "title": "Page",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "page_size",
      "required": false,
      "schema": {
       "default": 25,
       "minimum": 1,
       "title": "Page Size",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response List Users Api V1 Users Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "List Users",
    "tags": [
     "administration"
    ]
   },
   "post": {
    "description": "47.1.3 r3 \u2014 LegalMind never self-provisions an account.\n\nThe new account holds **no roles**. Authority is always a later, deliberate\ngrant, never a side effect of creation.",
    "operationId": "create_user_api_v1_users_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/UserCreate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Create User Api V1 Users Post",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Create User",
    "tags": [
     "administration"
    ]
   }
  },
  "/api/v1/users/{user_id}": {
   "delete": {
    "description": "Delete a user. Only DISABLED users can be deleted.",
    "operationId": "delete_user_api_v1_users__user_id__delete",
    "parameters": [
     {
      "in": "path",
      "name": "user_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "User Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Delete User Api V1 Users  User Id  Delete",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Delete User",
    "tags": [
     "administration"
    ]
   },
   "get": {
    "operationId": "get_user_api_v1_users__user_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "user_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "User Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Get User Api V1 Users  User Id  Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get User",
    "tags": [
     "administration"
    ]
   },
   "patch": {
    "description": "Editing an account is covered by S-9, not only granting to one.",
    "operationId": "update_user_api_v1_users__user_id__patch",
    "parameters": [
     {
      "in": "path",
      "name": "user_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "User Id",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/UserUpdate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Update User Api V1 Users  User Id  Patch",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Update User",
    "tags": [
     "administration"
    ]
   }
  },
  "/api/v1/users/{user_id}/roles": {
   "post": {
    "description": "Multi-role with union semantics (47.3 / SEC-03).\n\nThis is the mechanism locked Step 4 depends on: two users holding the same\nprimary role differ in legal authority because one *additionally holds* a\nLegal Decision Authority role.",
    "operationId": "grant_role_api_v1_users__user_id__roles_post",
    "parameters": [
     {
      "in": "path",
      "name": "user_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "User Id",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RoleGrant"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Grant Role Api V1 Users  User Id  Roles Post",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Grant Role",
    "tags": [
     "administration"
    ]
   }
  },
  "/api/v1/users/{user_id}/roles/{role_code}": {
   "delete": {
    "operationId": "revoke_role_api_v1_users__user_id__roles__role_code__delete",
    "parameters": [
     {
      "in": "path",
      "name": "user_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "User Id",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "role_code",
      "required": true,
      "schema": {
       "title": "Role Code",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Revoke Role Api V1 Users  User Id  Roles  Role Code  Delete",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Revoke Role",
    "tags": [
     "administration"
    ]
   }
  },
...
 }
}